requirements.txt
tensorflow
tensorflow[keras]
numpy
pillow
opencv-python